refactor: merge gl renderer into gl engine
