+++ packages/effects-core/src/render/renderer.ts
@@ -1,50 +1,78 @@
 import type { Matrix4, Vector3, Vector4 } from '@galacean/effects-math/es/core/index';
 import type { RendererComponent } from '../components';
 import type { Engine } from '../engine';
-import type { Material } from '../material';
-import type { LostHandler, RestoreHandler } from '../utils';
+import { Material } from '../material';
+import { sortByOrder } from '../utils';
 import type { FilterMode, Framebuffer, RenderTextureFormat } from './framebuffer';
-import type { Geometry } from './geometry';
-import type { RenderFrame, RenderingData } from './render-frame';
-import type { RenderPassClearAction } from './render-pass';
+import { Geometry } from './geometry';
+import type { RenderFrame } from './render-frame';
+import type { RenderPass, RenderPassClearAction } from './render-pass';
 import type { ShaderLibrary } from './shader';
-import { RenderTargetPool } from './render-target-pool';
 import type { Texture } from '../texture';
+import { math } from '..';
+import { glContext } from '../gl';
 
-export class Renderer implements LostHandler, RestoreHandler {
+// Blit shader 定义
+const BLIT_VERTEX_SHADER = `
+precision highp float;
+attribute vec2 aPos;
+varying vec2 vTex;
+void main(){
+    gl_Position = vec4(aPos, 0.0, 1.0);
+    vTex = (aPos + vec2(1.0)) / 2.0;
+}`;
+
+const BLIT_FRAGMENT_SHADER = `
+precision mediump float;
+varying vec2 vTex;
+uniform sampler2D _MainTex;
+void main(){
+    gl_FragColor = texture2D(_MainTex, vTex);
+}`;
+
+export class Renderer {
   static create: (engine: Engine) => Renderer;
 
   /**
   * 存放渲染需要用到的数据
   */
-  renderingData: RenderingData;
-  renderTargetPool: RenderTargetPool;
+  // renderingData: RenderingData;
   protected currentFramebuffer: Framebuffer | null = null;
+  protected disposed = false;
 
-  constructor (
-    public engine: Engine,
-  ) {
-    this.renderTargetPool = new RenderTargetPool(engine);
+  private blitGeometry: Geometry | null = null;
+  private blitMaterial: Material | null = null;
+
+  constructor (public engine: Engine) {
+  }
+
+  get renderingData () {
+    return this.engine.renderingData;
   }
 
   setGlobalFloat (name: string, value: number) {
-    // OVERRIDE
-  }
-
-  setGlobalInt (name: string, value: number) {
-    // OVERRIDE
+    this.checkGlobalUniform(name);
+    this.renderingData.currentFrame.globalUniforms.floats[name] = value;
   }
 
   setGlobalVector4 (name: string, value: Vector4) {
-    // OVERRIDE
+    this.checkGlobalUniform(name);
+    this.renderingData.currentFrame.globalUniforms.vector4s[name] = value;
   }
 
-  setGlobalVector3 (name: string, value: Vector3) {
-    // OVERRIDE
+  setGlobalInt (name: string, value: number) {
+    this.checkGlobalUniform(name);
+    this.renderingData.currentFrame.globalUniforms.ints[name] = value;
   }
 
   setGlobalMatrix (name: string, value: Matrix4) {
-    // OVERRIDE
+    this.checkGlobalUniform(name);
+    this.renderingData.currentFrame.globalUniforms.matrices[name] = value;
+  }
+
+  setGlobalVector3 (name: string, value: Vector3) {
+    this.checkGlobalUniform(name);
+    this.renderingData.currentFrame.globalUniforms.vector3s[name] = value;
   }
 
   getFramebuffer (): Framebuffer {
@@ -52,44 +80,31 @@ export class Renderer implements LostHandler, RestoreHandler {
   }
 
   setFramebuffer (framebuffer: Framebuffer | null) {
-    // OVERRIDE
+    if (framebuffer) {
+      this.currentFramebuffer = framebuffer;
+      this.currentFramebuffer.bind();
+      this.setViewport(framebuffer.viewport[0], framebuffer.viewport[1], framebuffer.viewport[2], framebuffer.viewport[3]);
+    } else {
+      this.currentFramebuffer = null;
+      this.engine.bindSystemFramebuffer();
+      this.setViewport(0, 0, this.getWidth(), this.getHeight());
+    }
   }
 
   setViewport (x: number, y: number, width: number, height: number) {
-    // OVERRIDE
-  }
-
-  resize (canvasWidth: number, canvasHeight: number) {
-    // OVERRIDE
+    this.engine.viewport(x, y, width, height);
   }
 
   clear (action: RenderPassClearAction) {
-    // OVERRIDE
+    this.engine.clear(action);
   }
 
   getWidth (): number {
-    // OVERRIDE
-    return 0;
+    return this.engine.getWidth();
   }
 
   getHeight (): number {
-    // OVERRIDE
-    return 0;
-  }
-
-  /**
-   * @override
-   * @param e
-   */
-  lost (e: Event): void {
-    // OVERRIDE
-  }
-
-  /**
-   * @override
-   */
-  restore (): void {
-    // OVERRIDE
+    return this.engine.getHeight();
   }
 
   /**
@@ -97,21 +112,64 @@ export class Renderer implements LostHandler, RestoreHandler {
    * @override
    * @returns
    */
-  getShaderLibrary (): ShaderLibrary | undefined {
-    // OVERRIDE
-    return undefined;
+  getShaderLibrary (): ShaderLibrary | null {
+    return this.engine.getShaderLibrary();
   }
 
   renderRenderFrame (renderFrame: RenderFrame) {
-    // OVERRIDE
+    const frame = renderFrame;
+    const passes = frame.renderPasses;
+
+    if (this.disposed) {
+      console.error('Renderer is destroyed, target: GLRenderer.');
+
+      return;
+    }
+
+    frame.renderer.engine.getShaderLibrary()?.compileAllShaders();
+    frame.setup();
+
+    this.setFramebuffer(null);
+
+    const currentCamera = frame.camera;
+
+    this.renderingData.currentFrame = frame;
+    this.renderingData.currentCamera = currentCamera;
+
+    this.setGlobalMatrix('effects_MatrixInvV', currentCamera.getInverseViewMatrix());
+    this.setGlobalMatrix('effects_MatrixV', currentCamera.getViewMatrix());
+    this.setGlobalMatrix('effects_MatrixVP', currentCamera.getViewProjectionMatrix());
+    this.setGlobalMatrix('_MatrixP', currentCamera.getProjectionMatrix());
+    this.setGlobalVector3('effects_WorldSpaceCameraPos', currentCamera.position);
+
+    // 根据 priority 排序 pass
+    sortByOrder(passes);
+
+    for (const pass of passes) {
+      this.renderRenderPass(pass);
+    }
+
+    for (const pass of passes) {
+      pass.onCameraCleanup(this);
+    }
+  }
+
+  renderRenderPass (pass: RenderPass): void {
+    this.renderingData.currentPass = pass;
+    // 配置当前 renderer 的 RT
+    pass.configure(this);
+    // 执行当前 pass
+    pass.execute(this);
   }
 
   renderMeshes (meshes: RendererComponent[]) {
-    // OVERRIDE
+    for (const mesh of meshes) {
+      mesh.render(this);
+    }
   }
 
   drawGeometry (geometry: Geometry, matrix: Matrix4, material: Material, subMeshIndex = 0) {
-    // OVERRIDE
+    this.engine.drawGeometry(geometry, matrix, material, subMeshIndex);
   }
 
   getTemporaryRT (
@@ -122,11 +180,11 @@ export class Renderer implements LostHandler, RestoreHandler {
     filter: FilterMode,
     format: RenderTextureFormat,
   ): Framebuffer {
-    return this.renderTargetPool.get(name, width, height, depthBuffer, filter, format);
+    return this.engine.renderTargetPool.get(name, width, height, depthBuffer, filter, format);
   }
 
   releaseTemporaryRT (rt: Framebuffer): void {
-    this.renderTargetPool.release(rt);
+    this.engine.renderTargetPool.release(rt);
   }
 
   /**
@@ -136,10 +194,78 @@ export class Renderer implements LostHandler, RestoreHandler {
    * @param material - 可选的自定义材质，不传则使用默认复制材质
    */
   blit (source: Texture, destination: Framebuffer | null, material?: Material): void {
-    // OVERRIDE
+    // 懒加载创建 blit geometry
+    if (!this.blitGeometry) {
+      this.blitGeometry = Geometry.create(this.engine, {
+        mode: glContext.TRIANGLE_STRIP,
+        attributes: {
+          aPos: {
+            type: glContext.FLOAT,
+            size: 2,
+            data: new Float32Array([-1, 1, -1, -1, 1, 1, 1, -1]),
+          },
+        },
+        drawCount: 4,
+      });
+    }
+
+    // 懒加载创建默认 blit material
+    if (!this.blitMaterial) {
+      this.blitMaterial = Material.create(this.engine, {
+        shader: {
+          vertex: BLIT_VERTEX_SHADER,
+          fragment: BLIT_FRAGMENT_SHADER,
+        },
+      });
+      this.blitMaterial.blending = false;
+      this.blitMaterial.depthTest = false;
+      this.blitMaterial.culling = false;
+    }
+
+    const blitMat = material || this.blitMaterial;
+
+    // 设置源纹理
+    blitMat.setTexture('_MainTex', source);
+
+    // 保存当前 framebuffer
+    const prevFramebuffer = this.currentFramebuffer;
+
+    // 设置目标
+    if (destination) {
+      const [x, y, width, height] = destination.viewport;
+
+      this.setFramebuffer(destination);
+      this.setViewport(x, y, width, height);
+    } else {
+      // 渲染到屏幕
+      this.setFramebuffer(null);
+      this.setViewport(0, 0, this.getWidth(), this.getHeight());
+    }
+
+    this.drawGeometry(this.blitGeometry, math.Matrix4.IDENTITY, blitMat);
+
+    // 恢复之前的 framebuffer
+    this.setFramebuffer(prevFramebuffer);
   }
 
   dispose (): void {
-    // OVERRIDE
+    if (this.disposed) {
+      return;
+    }
+
+    this.blitGeometry?.dispose();
+    this.blitGeometry = null;
+    this.blitMaterial?.dispose();
+    this.blitMaterial = null;
+
+    this.disposed = true;
+  }
+
+  private checkGlobalUniform (name: string) {
+    const globalUniforms = this.renderingData.currentFrame.globalUniforms;
+
+    if (!globalUniforms.uniforms.includes(name)) {
+      globalUniforms.uniforms.push(name);
+    }
   }
 }
