+++ packages/effects-core/src/scene-loader.ts
@@ -36,7 +36,7 @@ export class SceneLoader {
     const compileStart = performance.now();
 
     await new Promise(resolve => {
-      engine.renderer.getShaderLibrary()?.compileAllShaders(() => resolve(null));
+      engine.getShaderLibrary()?.compileAllShaders(() => resolve(null));
     });
 
     const compileTime = performance.now() - compileStart;
@@ -55,12 +55,8 @@ export class SceneLoader {
   }
 
   private static createComposition (scene: Scene, engine: Engine, options: SceneLoadOptions = {}): Composition {
-    const renderer = engine.renderer;
-    const composition = new Composition({
+    const composition = new Composition(engine, {
       ...options,
-      renderer,
-      width: renderer.getWidth(),
-      height: renderer.getHeight(),
       event: engine.eventSystem,
     }, scene);
