+++ packages/effects-webgl/src/gl-engine.ts
@@ -1,9 +1,16 @@
-import type { Composition, EngineOptions, Nullable, Texture, Texture2DSourceOptionsVideo, math } from '@galacean/effects-core';
-import { Engine, GPUCapability, SceneLoader, assertExist, glContext, isIOS } from '@galacean/effects-core';
-import { GLRenderer } from './gl-renderer';
+import type { Composition, EngineOptions, Geometry, Material, Nullable, RenderPassClearAction, ShaderLibrary, Texture, Texture2DSourceOptionsVideo, math } from '@galacean/effects-core';
+import { Engine, GPUCapability, Renderer, SceneLoader, TextureLoadAction, assertExist, glContext, isIOS, logger } from '@galacean/effects-core';
 import { GLShaderLibrary } from './gl-shader-library';
 import type { GLTexture } from './gl-texture';
 import { GLContextManager } from './gl-context-manager';
+import { assignInspectorName } from './gl-renderer-internal';
+import type { GLFramebuffer } from './gl-framebuffer';
+import type { GLGPUBuffer } from './gl-gpu-buffer';
+import type { GLRenderbuffer } from './gl-renderbuffer';
+import { GLVertexArrayObject } from './gl-vertex-array-object';
+import type { GLGeometry } from './gl-geometry';
+import type { GLMaterial } from './gl-material';
+import type { GLShaderVariant } from './gl-shader';
 
 type Color = math.Color;
 type Vector2 = math.Vector2;
@@ -47,14 +54,25 @@ export class GLEngine extends Engine {
         this.ticker?.pause();
         this.restoreCompositionsCache = this.compositions.slice();
         this.compositions.forEach(comp => comp.lost(e));
-        this.renderer.lost(e);
+        e.preventDefault();
+        logger.error(`WebGL context lost. Event target: ${e.target}.`);
         this.emit('contextlost', { engine: this, e });
       },
     });
 
     this.context.addRestoreHandler({
       restore: async () => {
-        this.renderer.restore();
+        // FIXME: 需要测试下lost和restore流程
+        const { gl } = this.context;
+
+        if (!gl) {
+          throw new Error('Can not restore automatically because losing gl context.');
+        }
+
+        this.reset();
+        this.shaderLibrary = new GLShaderLibrary(this);
+        this.gpuCapability = new GPUCapability(gl);
+
         await Promise.all(this.restoreCompositionsCache.map(async composition => {
           const { time: currentTime, url, speed, reusable, renderOrder, transform, videoState } = composition;
           const newComposition = await SceneLoader.load(url, this);
@@ -102,11 +120,166 @@ export class GLEngine extends Engine {
     this.reset();
     this.gpuCapability = new GPUCapability(gl);
     this.shaderLibrary = new GLShaderLibrary(this);
-    this.renderer = new GLRenderer(this);
+    this.renderer = new Renderer(this);
     this.maxTextureCount = this.gl.TEXTURE0 + this.gl.getParameter(this.gl.MAX_COMBINED_TEXTURE_IMAGE_UNITS) - 1;
 
     // resize need gl renderer initialized
     this.resize();
+  }
+
+  override getWidth (): number {
+    return this.gl.drawingBufferWidth;
+  }
+
+  override getHeight (): number {
+    return this.gl.drawingBufferHeight;
+  }
+
+  override getShaderLibrary (): ShaderLibrary | null {
+    return this.shaderLibrary;
+  }
+
+  createGLFramebuffer (name?: string): WebGLFramebuffer | null {
+    const fbo = this.gl.createFramebuffer();
+
+    if (fbo) {
+      assignInspectorName(fbo, name, name);
+    } else {
+      throw new Error(`Failed to create WebGL framebuffer. gl isContextLost=${this.gl.isContextLost()}`);
+    }
+
+    return fbo;
+  }
+
+  createVAO (name?: string): GLVertexArrayObject | undefined {
+    const ret = new GLVertexArrayObject(this, name);
+
+    return ret;
+  }
+
+  deleteGLTexture (texture: GLTexture) {
+    if (texture.textureBuffer && !this.disposed) {
+      this.gl.deleteTexture(texture.textureBuffer);
+      texture.textureBuffer = null;
+    }
+  }
+
+  deleteGPUBuffer (buffer: GLGPUBuffer | null) {
+    if (buffer && !this.disposed) {
+      this.gl.deleteBuffer(buffer.glBuffer);
+      // @ts-expect-error
+      delete buffer.glBuffer;
+    }
+  }
+
+  deleteGLFramebuffer (framebuffer: GLFramebuffer) {
+    if (framebuffer && !this.disposed) {
+      this.gl.deleteFramebuffer(framebuffer.fbo as WebGLFramebuffer);
+      delete framebuffer.fbo;
+    }
+  }
+
+  deleteGLRenderbuffer (renderbuffer: GLRenderbuffer) {
+    if (renderbuffer && !this.disposed) {
+      this.gl.deleteRenderbuffer(renderbuffer.buffer);
+      renderbuffer.buffer = null;
+    }
+  }
+
+  override drawGeometry (geometry: Geometry, matrix: Matrix4, material: Material, subMeshIndex = 0): void {
+    if (!geometry || !material) {
+      return;
+    }
+
+    material.initialize();
+    geometry.initialize();
+    geometry.flush();
+    const renderingData = this.renderingData;
+
+    material.setMatrix('effects_ObjectToWorld', matrix);
+
+    try {
+      material.use(this.renderer, renderingData.currentFrame.globalUniforms);
+    } catch (e) {
+      console.error(e);
+
+      this.renderErrors.add(e as Error);
+
+      return;
+    }
+
+    const gl = this.gl;
+
+    if (!gl) {
+      console.warn('GLGPURenderer has not bound a gl object, unable to render geometry.');
+
+      return;
+    }
+
+    const glGeometry = geometry as GLGeometry;
+    const glMaterial = material as GLMaterial;
+    const program = (glMaterial.shaderVariant as GLShaderVariant).program;
+
+    if (!program) {
+      return;
+    }
+
+    const vao = program.setupAttributes(glGeometry);
+    const indicesBuffer = glGeometry.indicesBuffer;
+    let offset = glGeometry.drawStart;
+    let count = glGeometry.drawCount;
+    const mode = glGeometry.mode;
+    const subMeshes = glGeometry.subMeshes;
+
+    if (subMeshes && subMeshes.length) {
+      const subMesh = subMeshes[subMeshIndex];
+
+      // FIXME: 临时处理3D线框状态下隐藏模型
+      if (count < 0) {
+        return;
+      }
+      offset = subMesh.offset;
+      if (indicesBuffer) {
+        count = subMesh.indexCount ?? 0;
+      } else {
+        count = subMesh.vertexCount;
+      }
+    }
+    if (indicesBuffer) {
+      gl.drawElements(mode, count, indicesBuffer.type, offset ?? 0);
+    } else {
+      gl.drawArrays(mode, offset, count);
+    }
+    vao?.unbind();
+  }
+
+  override clear (action: RenderPassClearAction): void {
+    let bit = 0;
+
+    if (action.colorAction === TextureLoadAction.clear) {
+      const clearColor = action.clearColor;
+
+      if (clearColor) {
+        this.clearColor(clearColor[0], clearColor[1], clearColor[2], clearColor[3]);
+      }
+      this.colorMask(true, true, true, true);
+      bit = glContext.COLOR_BUFFER_BIT;
+    }
+    if (action.stencilAction === TextureLoadAction.clear) {
+      this.stencilMask(0xff);
+      this.clearStencil(action.clearStencil || 0);
+      bit = bit | glContext.STENCIL_BUFFER_BIT;
+    }
+    if (action.depthAction === TextureLoadAction.clear) {
+      const depth = action.clearDepth as number;
+
+      this.depthMask(true);
+      this.clearDepth(Number.isFinite(depth) ? depth : 1);
+      bit = bit | glContext.DEPTH_BUFFER_BIT;
+    }
+    if (bit) {
+      this.gl.clear(bit);
+    }
   }
 
   override dispose () {
@@ -116,6 +289,7 @@ export class GLEngine extends Engine {
     super.dispose();
 
     this.renderer.dispose();
+    this.renderTargetPool.dispose();
     this.shaderLibrary?.dispose();
     this.context.dispose();
     this.reset();
@@ -193,7 +367,7 @@ export class GLEngine extends Engine {
   /**
    * 绑定系统 framebuffer
    */
-  bindSystemFramebuffer () {
+  override bindSystemFramebuffer () {
     this.bindFramebuffer(this.gl.FRAMEBUFFER, null);
   }
 
@@ -206,17 +380,6 @@ export class GLEngine extends Engine {
    */
   useProgram (program: WebGLProgram | null) {
     this.set1('useProgram', program);
-  }
-
-  /**
-   * 使用预设值来清空缓冲
-   * @param mask
-   * example:
-   * gl.clear(gl.DEPTH_BUFFER_BIT);
-   * gl.clear(gl.DEPTH_BUFFER_BIT | gl.COLOR_BUFFER_BIT);
-   */
-  clear (mask: number) {
-    this.gl.clear(mask);
   }
 
   /*** depth start ***/
@@ -506,7 +669,7 @@ export class GLEngine extends Engine {
    * example:
    * gl.viewport(0, 0, width, height);
    */
-  viewport (x: number, y: number, width: number, height: number) {
+  override viewport (x: number, y: number, width: number, height: number) {
     this.set4('viewport', x, y, width, height);
   }
