+++ packages/effects-webgl/src/gl-renderbuffer.ts
@@ -1,17 +1,16 @@
-import type { RenderbufferProps } from '@galacean/effects-core';
+import type { RenderbufferProps, Renderer } from '@galacean/effects-core';
 import { throwDestroyedError, Renderbuffer, logger } from '@galacean/effects-core';
-import type { GLRenderer } from './gl-renderer';
 import type { GLEngine } from './gl-engine';
 
 export class GLRenderbuffer extends Renderbuffer {
   buffer: WebGLRenderbuffer | null;
 
   private initialized = false;
-  private renderer?: GLRenderer | null;
+  private renderer?: Renderer | null;
 
   constructor (
     props: RenderbufferProps,
-    renderer?: GLRenderer,
+    renderer?: Renderer,
   ) {
     super(props);
 
@@ -20,14 +19,14 @@ export class GLRenderbuffer extends Renderbuffer {
     }
   }
 
-  initialize (renderer: GLRenderer) {
+  initialize (renderer: Renderer) {
     if (this.initialized) {
       return;
     }
 
     this.initialized = true;
     this.renderer = renderer;
-    this.buffer = renderer.createGLRenderbuffer(this) as WebGLRenderbuffer;
+    this.buffer = (renderer.engine as GLEngine).gl.createRenderbuffer() as WebGLRenderbuffer;
   }
 
   setSize (width: number, height: number) {
@@ -42,9 +41,10 @@ export class GLRenderbuffer extends Renderbuffer {
     }
 
     if (width !== this.size[0] || height !== this.size[1]) {
-      const { gl, engine: state } = this.renderer;
+      const engine = this.renderer.engine as GLEngine;
+      const gl = engine.gl;
 
-      (state as GLEngine).bindRenderbuffer(gl.RENDERBUFFER, this.buffer);
+      engine.bindRenderbuffer(gl.RENDERBUFFER, this.buffer);
       if (width && height) {
         gl.renderbufferStorage(gl.RENDERBUFFER, this.format, this.size[0] = width, this.size[1] = height);
       } else {
@@ -55,7 +55,7 @@ export class GLRenderbuffer extends Renderbuffer {
 
   dispose () {
     if (this.renderer) {
-      this.renderer.deleteGLRenderbuffer(this);
+      (this.renderer.engine as GLEngine).deleteGLRenderbuffer(this);
       this.renderer = null;
       this.buffer = null;
     }
