+++ plugin-packages/stats/src/stats-component.ts
@@ -1,19 +1,19 @@
-import type { GLEngine, GLRenderer } from '@galacean/effects';
-import { Behaviour } from '@galacean/effects';
+import type { GLEngine } from '@galacean/effects';
+import { Component } from '@galacean/effects';
 import { Monitor } from './monitor';
 import type { StatsOptions } from './stats';
 
 /**
  * Stats 组件，用于监控渲染性能
  */
-export class StatsComponent extends Behaviour {
+export class StatsComponent extends Component {
   /**
    * 监控对象
    */
   monitor: Monitor;
 
   init (options: Required<StatsOptions>): void {
-    const renderer = this.engine.renderer as GLRenderer;
+    const renderer = this.engine.renderer;
     const gl = (renderer.engine as GLEngine).gl;
 
     this.monitor = new Monitor(gl, options);
