+++ web-packages/test/unit/src/effects-webgl/gl-geometry.spec.ts
@@ -30,7 +30,7 @@ describe('webgl/gl-geometry', () => {
     canvas = document.createElement('canvas');
     engine = new GLEngine(canvas, { glType: 'webgl2' });
     renderer = engine.renderer;
-    glRenderer = renderer;
+    glRenderer = engine;
     gl = glRenderer.gl;
     geometry = new GLGeometry(engine, option);
   });
@@ -537,8 +537,8 @@ describe('webgl/gl-geometry', () => {
 
   // drawCount小于0时不会触发drawCall
   it('geometry with drawCount < 0 would not trigger draw call', function () {
-    const ret = createMesh(glRenderer, -2);
-    const pass = new RenderPass(glRenderer);
+    const ret = createMesh(renderer, -2);
+    const pass = new RenderPass(renderer);
 
     pass.addMesh(ret.mesh);
     ret.geom.drawCount = -1;
@@ -559,7 +559,7 @@ describe('webgl/gl-geometry', () => {
   // drawCount等于0时不会触发drawCall
   it('geometry with drawCount == 0 would not invoke draw call', function () {
     const ret = createMesh(renderer, 0);
-    const pass = new RenderPass(glRenderer);
+    const pass = new RenderPass(renderer);
 
     pass.addMesh(ret.mesh);
     const frame = new RenderFrame({
